requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
Deprecated==1.2.14
dj-database-url==2.1.0
Django==5.0.2
django-configurations==2.5
django-stubs==4.2.7
django-stubs-ext==4.2.7
idna==3.6
psycopg==3.1.18
psycopg-binary==3.1.18
pycparser==2.21
PyGithub==2.2.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.31.0
sqlparse==0.4.4
types-pytz==2024.1.0.20240203
types-PyYAML==6.0.12.12
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
